PDJB-765: Extract base pages for electrical safety page objects
