+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/integration/pageObjects/pages/updateElectricalSafetyJourneyPages/CheckElectricalCertUploadsFormPageUpdateElectricalSafety.kt
@@ -2,10 +2,9 @@ package uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.updateElec
 
 import com.microsoft.playwright.Page
 import uk.gov.communities.prsdb.webapp.controllers.UpdateElectricalSafetyController
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.PostForm
-import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.SecondaryButton
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.components.Table
 import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.basePages.BasePage
+import uk.gov.communities.prsdb.webapp.integration.pageObjects.pages.propertyRegistrationJourneyPages.CheckElectricalCertUploadsFormPagePropertyRegistration.CheckUploadsForm
 import uk.gov.communities.prsdb.webapp.journeys.propertyRegistration.steps.CheckElectricalCertUploadsStep
 
 class CheckElectricalCertUploadsFormPageUpdateElectricalSafety(
@@ -17,16 +16,6 @@ class CheckElectricalCertUploadsFormPageUpdateElectricalSafety(
             .replace("{propertyOwnershipId}", urlArguments["propertyOwnershipId"]!!) +
             "/${CheckElectricalCertUploadsStep.ROUTE_SEGMENT}",
     ) {
+    val table = Table(page)
     val form = CheckUploadsForm(page)
-    val table = CheckUploadsTable(page)
-
-    class CheckUploadsForm(
-        page: Page,
-    ) : PostForm(page) {
-        val addAnotherButton = SecondaryButton(locator)
-    }
-
-    class CheckUploadsTable(
-        page: Page,
-    ) : Table(page)
 }
